requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1.1
django-environ==0.11.2
django-allauth==65.0.2
django-extensions==3.2.3
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
python-decouple==3.8
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
